+++ app-Prefs/AppDelegate.swift
@@ -107,88 +107,20 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             let alertC = UIAlertController(title: SwitchLanguageTool.getLocalString(of: "Warning"), message: SwitchLanguageTool.getLocalString(of: "importWarning"), preferredStyle: .alert)
             let cancelAction = UIAlertAction(title: SwitchLanguageTool.getLocalString(of: "Cancel"), style: .cancel, handler: { (_) in
             })
-            let sureAction = UIAlertAction(title: SwitchLanguageTool.getLocalString(of: "Sure"), style: .destructive, handler: { (_) in
-                
-                if url.absoluteString.hasSuffix(".plist") {
-                    if let dict = NSDictionary(contentsOf: url) as? Dictionary<String, AnyObject> {
-                        if dict.keys.contains("name") && dict.keys.contains("settings") {
-                            if (dict["name"] as! String == "app-Prefs") && (dict["settings"]?.isKind(of: NSArray.self))! {
-                                let tempArr = dict["settings"] as! NSArray
-                                var ret = true
-                                for tempDict in tempArr {
-                                    if (tempDict as AnyObject).isKind(of: NSDictionary.self) {
-                                        if ((tempDict as! NSDictionary).allKeys as NSArray).contains("action") && ((tempDict as! NSDictionary).allKeys as NSArray).contains("name") && ((tempDict as! NSDictionary).allKeys as NSArray).contains("isDeleted") && ((tempDict as! NSDictionary).allKeys as NSArray).contains("sortNum") && ((tempDict as! NSDictionary).allKeys as NSArray).contains("type") {
-                                            if ((tempDict as! NSDictionary)["action"] as AnyObject).isKind(of: NSString.self) && ((tempDict as! NSDictionary)["name"] as AnyObject).isKind(of: NSString.self) && ((tempDict as! NSDictionary)["type"] as AnyObject).isKind(of: NSString.self) && ((tempDict as! NSDictionary)["sortNum"] as AnyObject).isKind(of: NSNumber.self) && (((tempDict as! NSDictionary)["isDeleted"] as? Bool == false) || ((tempDict as! NSDictionary)["isDeleted"] as? Bool == true))  {
-                                                
-                                            } else {
-                                                ret = false
-                                            }
-                                        } else {
-                                            ret = false
-                                        }
-                                    } else {
-                                        ret = false
-                                    }
-                                }
-                                if ret != true {
-                                    self.alertWrongFormat()
-                                    return
-                                }
-                                let container = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.chengluffy.app-Prefs")
-                                let realmURL = container!.appendingPathComponent("defualt.realm")
-                                Realm.Configuration.defaultConfiguration.fileURL = realmURL
-                                do {
-                                    let realm = try! Realm()
-                                    try realm.write {
-                                        realm.deleteAll()
-                                    }
-                                    try realm.write {
-                                        let arr = dict["settings"] as! Array<Dictionary<String, Any>>
-                                        for subDict in arr {
-                                            let model = Setting()
-                                            model.action = subDict["action"] as! String
-                                            model.name = subDict["name"] as! String
-                                            model.isDeleted = subDict["isDeleted"] as! Bool
-                                            model.sortNum = subDict["sortNum"] as! NSNumber
-                                            model.type = subDict["type"] as! String
-                                            print(model.name)
-                                            realm.add(model)
-                                        }
-                                    }
-                                    SVProgressHUD.showSuccess(withStatus: SwitchLanguageTool.getLocalString(of: "Success!"))
-                                } catch {
-                                    print(error)
-                                    self.alertWrongFormat()
-                                }
-                                
-                                ((self.window?.rootViewController as! UINavigationController).viewControllers.first as! ViewController).refresh()
-                                try! FileManager.default.removeItem(at: url)
-                            } else {
-                                print("name' value isn't app-Prefs")
-                                self.alertWrongFormat()
-                                try! FileManager.default.removeItem(at: url)
-                            }
-                        } else {
-                            print("no key: name")
-                            self.alertWrongFormat()
-                            try! FileManager.default.removeItem(at: url)
-                        }
-                        
-                    } else {
-                        self.alertWrongFormat()
-                        try! FileManager.default.removeItem(at: url)
-                    }
-                } else {
-                    print("not a plist file")
-                    self.alertWrongFormat()
-                    try! FileManager.default.removeItem(at: url)
-                }
+            let addAllAction = UIAlertAction(title: SwitchLanguageTool.getLocalString(of: "Delete Local & Import All"), style: .destructive, handler: { (_) in
+                let _ = ConfigTool.import(from: url, deleteAll: true)
+            })
+            let addNotExsitAction = UIAlertAction(title: SwitchLanguageTool.getLocalString(of: "Only import Not Exsit"), style: .default, handler: { (_) in
+                let _ = ConfigTool.import(from: url, deleteAll: false)
+            })
+            alertC.addAction(cancelAction)
+            alertC.addAction(addAllAction)
+            alertC.addAction(addNotExsitAction)
+            
+            (self.window?.rootViewController as! UINavigationController).popToRootViewController(animated: true)
+            window?.rootViewController?.present(alertC, animated: true, completion: { 
             })
             
-            alertC.addAction(cancelAction)
-            alertC.addAction(sureAction)
-            
-            window?.rootViewController?.present(alertC, animated: true, completion: nil)
         } else if url.absoluteString.hasPrefix("FastOpenJSON://") {
             let textVC = TextViewController()
             textVC.urlStr = url
@@ -197,10 +129,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         
         
         return true
-    }
-    
-    func alertWrongFormat() {
-        SVProgressHUD.showError(withStatus: SwitchLanguageTool.getLocalString(of: "WrongFormat"))
     }
 
 }