+++ ChangeLog.md
@@ -1,3 +1,13 @@
+build version: 2017030500
+---
+- support select type when Enter your action.
+
+build version: 2017030318
+---
+- import but not delete local.
+- update treeview.
+- export & import use json file.
+
 build version: 2017022822
 ---
 - really fixed Today Widget layout error.